Create LICENSE.txt
This is the same license used in some other Enthought projects.  The BSD license requires all copies of the source code to include the license text.  Further, without a license file it is unclear which of the various BSD licenses this project falls under.  I am assuming here it would be the same across Enthought projects.